+++ cl_dll/studio_util.cpp
@@ -189,10 +189,10 @@ void QuaternionSlerp(vec4_t p, vec4_t q, float t, vec4_t qt)
 	{
 		if ((1.0 - cosom) > 0.000001)
 		{
-			omega = acos(cosom);
-			sinom = sin(omega);
-			sclp = sin((1.0 - t) * omega) / sinom;
-			sclq = sin(t * omega) / sinom;
+			omega = acosf(cosom);
+			sinom = sinf(omega);
+			sclp = sinf((1.0f - t) * omega) / sinom;
+			sclq = sinf(t * omega) / sinom;
 		}
 		else
 		{
